Pre: Add Gradle TestKit test.
For things that are interacting with the Gradle model in non-trivial
ways, Gradle TestKit (and, often, special output tasks like `doTest`
here) are a good way to test things: better than unit tests, which
have to try to arrange for a dummy `Project`, etc.

I have little success running these tests in Android Studio, for
reasons I've never understood.  `./gradlew gradle-plugin:test` works
just fine.  They are slow because an internal Gradle is invoked for
each test.

It's awkward to use temporary folders with JUnit 5 before the 5.4
series so I updated the dependency.