Part 2: Add support for consuming a local GeckoView and a sample app.
There are many notes captured in the plugin source.  It's **very**
hard to test against the Android plugin because there's no uniform way
to configure the Android SDK under the hood.  Instead, I've provided a
sample app... which also shows how to work on the plugin using Gradle
composite builds.

This is working well against firefox-tv and fenix.